refactor conversation input actions and simplify loading state
