+++ x-pack/platform/plugins/shared/onechat/public/application/components/conversations/conversation_input/conversation_input_actions.tsx
@@ -5,93 +5,36 @@
  * 2.0.
  */
 
-import { EuiButtonIcon, EuiFlexGroup, EuiFlexItem, useEuiTheme } from '@elastic/eui';
-import { i18n } from '@kbn/i18n';
+import { EuiFlexGroup, EuiFlexItem } from '@elastic/eui';
 import React from 'react';
-import { css } from '@emotion/react';
-import type { AgentDefinition } from '@kbn/onechat-common';
-import { useConversationActions } from '../../../hooks/use_conversation_actions';
-import { AgentDisplay } from '../agent_display';
-import { AgentSelectDropdown } from '../agent_select_dropdown';
-import { useSendMessage } from '../../../context/send_message/send_message_context';
+import { ConversationAgentSelector } from '../conversation_agent_selector';
+import { ConversationActionButton } from './conversation_action_button';
+
 interface ConversationInputActionsProps {
   onSubmit: () => void;
   isSubmitDisabled: boolean;
   resetToPendingMessage: () => void;
-  agents?: AgentDefinition[];
-  currentAgent?: AgentDefinition | null;
-  isLoading?: boolean;
-  hasActiveConversation?: boolean;
+  agentId?: string;
 }
-
-const labels = {
-  cancel: i18n.translate('xpack.onechat.conversationInputForm.cancel', {
-    defaultMessage: 'Cancel',
-  }),
-  submit: i18n.translate('xpack.onechat.conversationInputForm.submit', {
-    defaultMessage: 'Submit',
-  }),
-};
 
 export const ConversationInputActions: React.FC<ConversationInputActionsProps> = ({
   onSubmit,
   isSubmitDisabled,
   resetToPendingMessage,
-  agents,
-  currentAgent,
-  isLoading,
-  hasActiveConversation,
+  agentId,
 }) => {
-  const { setAgentId } = useConversationActions();
-  const { canCancel, cancel } = useSendMessage();
-  const { euiTheme } = useEuiTheme();
-  const cancelButtonStyles = css`
-    background-color: ${euiTheme.colors.backgroundLightText};
-  `;
   return (
     <EuiFlexItem grow={false}>
       <EuiFlexGroup gutterSize="s" responsive={false} alignItems="center" justifyContent="flexEnd">
         <EuiFlexItem grow={false}>
-          {hasActiveConversation ? (
-            <AgentDisplay currentAgent={currentAgent} isLoading={isLoading} />
-          ) : (
-            <AgentSelectDropdown
-              selectedAgentId={currentAgent?.id}
-              onAgentChange={(newAgentId: string) => {
-                setAgentId(newAgentId);
-              }}
-              agents={agents}
-              isLoading={isLoading}
-            />
-          )}
+          <ConversationAgentSelector agentId={agentId} />
         </EuiFlexItem>
         <EuiFlexItem grow={false}>
-          {canCancel ? (
-            <EuiButtonIcon
-              aria-label={labels.cancel}
-              data-test-subj="onechatAppConversationInputFormCancelButton"
-              iconType="stopFilled"
-              size="m"
-              color="text"
-              css={cancelButtonStyles}
-              onClick={() => {
-                if (canCancel) {
-                  cancel();
-                  resetToPendingMessage();
-                }
-              }}
-            />
-          ) : (
-            <EuiButtonIcon
-              aria-label={labels.submit}
-              data-test-subj="onechatAppConversationInputFormSubmitButton"
-              iconType="sortUp"
-              display="fill"
-              size="m"
-              disabled={isSubmitDisabled}
-              onClick={onSubmit}
-            />
-          )}
+          <ConversationActionButton
+            onSubmit={onSubmit}
+            isSubmitDisabled={isSubmitDisabled}
+            resetToPendingMessage={resetToPendingMessage}
+          />
         </EuiFlexItem>
       </EuiFlexGroup>
     </EuiFlexItem>
