+++ x-pack/platform/plugins/shared/onechat/public/application/components/conversations/conversation_input/conversation_input_form.tsx
@@ -12,7 +12,7 @@ import React, { useState } from 'react';
 import { useSendMessage } from '../../../context/send_message/send_message_context';
 import { useIsSendingMessage } from '../../../hooks/use_is_sending_message';
 import { ConversationInputTextArea } from './conversation_input_text_area';
-import { useAgentId, useHasActiveConversation } from '../../../hooks/use_conversation';
+import { useAgentId } from '../../../hooks/use_conversation';
 import { useOnechatAgents } from '../../../hooks/agents/use_agents';
 import { ConversationInputActions } from './conversation_input_actions';
 
@@ -29,9 +29,8 @@ export const ConversationInputForm: React.FC<ConversationInputFormProps> = ({ on
   const [input, setInput] = useState('');
   const { sendMessage, pendingMessage } = useSendMessage();
   const { euiTheme } = useEuiTheme();
-  const { agents, isLoading, isFetched } = useOnechatAgents();
+  const { agents, isFetched } = useOnechatAgents();
   const agentId = useAgentId();
-  const hasActiveConversation = useHasActiveConversation();
 
   // Find the current agent from the agents array
   const currentAgent = agents?.find((agent) => agent.id === agentId);
@@ -91,10 +90,7 @@ export const ConversationInputForm: React.FC<ConversationInputFormProps> = ({ on
               setInput(pendingMessage);
             }
           }}
-          agents={agents}
-          currentAgent={currentAgent}
-          isLoading={isLoading}
-          hasActiveConversation={hasActiveConversation}
+          agentId={agentId}
         />
       )}
     </EuiFlexGroup>
